Refactor
I put the quest and other stuff in the Inventory script. Not sure if I will refactor it again but for now, everything can have access to the GameManager and from it you can access functions like DisplayQuests... These functions then don't have the logic but rather call the function with the logic in the Inventory script. This is so the Inventory is not accessible. Not sure if the double function thing is good tho. I maybe make it so that the Inventory is accessible too and you can just access the function directly or if I have it like this and make it that not everything can call those functions. Not sure tho